fix starknet-js v3 version to 3.18.2
