+++ burner/extension/package.json
@@ -36,7 +36,7 @@
     "@types/elliptic": "^6.4.14",
     "bn.js": "^5.2.1",
     "jest": "^28.1.3",
-    "starknet3x": "npm:starknet@3.18",
+    "starknet3x": "npm:starknet@3.18.2",
     "starknet4x": "npm:starknet@^4.4.2",
     "ts-jest": "^28.0.8",
     "ts-node": "^10.9.1",
