+++ burner/keyring/lib/3.x/account.js
@@ -1,68 +1,10 @@
-import { Account, Signer, Provider, ec, transaction } from "starknet";
+import { Account, Signer, Provider, ec } from "starknet3";
 import { notify } from "../shared/message";
 import { getLocalStorage } from "lib/storage";
 import { newLog } from "lib/shared/log";
-import { SessionAccount } from "@argent/x-sessions";
+import { account4sessionEventHandler } from "./account4session";
 
 const log = newLog();
-
-const estimateFee = async (account, data, key) => {
-  {
-    const { calls, estimateFeeDetails } = data;
-    let estimateFeeResponse;
-    try {
-      estimateFeeResponse = await account.estimateFee(
-        calls,
-        estimateFeeDetails
-      );
-    } catch (e) {
-      return notify({
-        type: "account3x_EstimateFeeResponse",
-        key,
-        exception: e.toString(),
-      });
-    }
-    const { overall_fee, gas_consumed, gas_price, suggestedMaxFee } =
-      estimateFeeResponse;
-    return notify({
-      type: "account3x_EstimateFeeResponse",
-      data: {
-        overall_fee: overall_fee.toString("hex"),
-        gas_consumed: gas_consumed.toString("hex"),
-        gas_price: gas_price.toString("hex"),
-        suggestedMaxFee: suggestedMaxFee.toString("hex"),
-      },
-      key,
-    });
-  }
-};
-
-const execute = async (account, data, key) => {
-  console.log("yes");
-  const { transactions, abis, transactionsDetail } = data;
-  console.log("yes 1", transactions, abis, transactionsDetail);
-  let executeResponse;
-  try {
-    executeResponse = await account.execute(
-      transactions,
-      abis,
-      transactionsDetail
-    );
-  } catch (e) {
-    console.log("yes 2", e.toString());
-    return notify({
-      type: "account3x_ExecuteResponse",
-      key,
-      exception: e.toString(),
-    });
-  }
-  console.log("yes notify", executeResponse);
-  return notify({
-    type: "account3x_ExecuteResponse",
-    data: executeResponse,
-    key,
-  });
-};
 
 const signMessage = async (account, data, key) => {
   const typedData = data;
@@ -130,6 +72,12 @@ const getNonce = async (account, data, key) => {
 
 export const accountEventHandler = async (type, data, key) => {
   log.debug(type, key, data);
+  if (type === "account3x_EstimateFee" || type === "account3x_Execute") {
+    // this is rerouted to a function that relies on starknet-js v4 because it
+    // is needed yet, most people developing with v3, including starknet-react
+    // we want to remain with the v3 implementation for a a few more days...
+    return await account4sessionEventHandler(type, data, key);
+  }
   const sessionKey = getLocalStorage("bwsessionkey");
   const token = getLocalStorage("bwsessiontoken");
   const parsedToken = JSON.parse(token);
@@ -139,17 +87,7 @@ export const accountEventHandler = async (type, data, key) => {
   const signer = new Signer(keypair);
   const provider = new Provider({ sequencer: { network: "alpha-goerli" } });
   const account = new Account(provider, address, signer);
-  const sessionAccount = new SessionAccount(
-    provider,
-    address,
-    signer,
-    parsedToken
-  );
   switch (type) {
-    case "account3x_EstimateFee":
-      return await estimateFee(sessionAccount, data, key);
-    case "account3x_Execute":
-      return await execute(sessionAccount, data, key, parsedToken);
     case "account3x_SignMessage":
       return await signMessage(account, data, key);
     case "account3x_HashMessage":
