+++ drone/package.json
@@ -1,41 +1,41 @@
 {
-	"name": "drone",
-	"version": "0.0.1",
-	"scripts": {
-		"dev": "vite dev",
-		"build": "vite build",
-		"package": "svelte-kit package",
-		"preview": "vite preview",
-		"prepare": "svelte-kit sync",
-		"test": "playwright test",
-		"check": "svelte-check --tsconfig ./tsconfig.json",
-		"check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
-		"lint": "prettier --check --plugin-search-dir=. . && eslint .",
-		"format": "prettier --write --plugin-search-dir=. ."
-	},
-	"devDependencies": {
-		"@argent/x-sessions": "^4.8.0",
-		"@playwright/test": "^1.25.2",
-		"@sveltejs/adapter-auto": "^1.0.0-next.72",
-		"@sveltejs/kit": "^1.0.0-next.479",
-		"@types/qrcode": "^1.4.2",
-		"@typescript-eslint/eslint-plugin": "^5.31.0",
-		"@typescript-eslint/parser": "^5.31.0",
-		"buffer": "^6.0.3",
-		"eslint": "^8.20.0",
-		"eslint-config-prettier": "^8.3.0",
-		"eslint-plugin-svelte3": "^4.0.0",
-		"get-starknet": "^1.4.2",
-		"prettier": "^2.6.2",
-		"prettier-plugin-svelte": "^2.7.0",
-		"qrcode": "^1.5.1",
-		"starknet": "^3.18.2",
-		"svelte": "^3.50.1",
-		"svelte-check": "^2.7.1",
-		"svelte-preprocess": "^4.10.6",
-		"tslib": "^2.3.1",
-		"typescript": "^4.8.3",
-		"vite": "^3.0.4"
-	},
-	"type": "module"
+  "name": "drone",
+  "version": "0.0.1",
+  "scripts": {
+    "dev": "vite dev",
+    "build": "vite build",
+    "package": "svelte-kit package",
+    "preview": "vite preview",
+    "prepare": "svelte-kit sync",
+    "test": "playwright test",
+    "check": "svelte-check --tsconfig ./tsconfig.json",
+    "check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
+    "lint": "prettier --check --plugin-search-dir=. . && eslint .",
+    "format": "prettier --write --plugin-search-dir=. ."
+  },
+  "devDependencies": {
+    "@argent/x-sessions": "^4.8.0",
+    "@playwright/test": "^1.25.2",
+    "@sveltejs/adapter-auto": "^1.0.0-next.72",
+    "@sveltejs/kit": "^1.0.0-next.479",
+    "@types/qrcode": "^1.4.2",
+    "@typescript-eslint/eslint-plugin": "^5.31.0",
+    "@typescript-eslint/parser": "^5.31.0",
+    "buffer": "^6.0.3",
+    "eslint": "^8.20.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-svelte3": "^4.0.0",
+    "get-starknet": "^1.4.2",
+    "prettier": "^2.6.2",
+    "prettier-plugin-svelte": "^2.7.0",
+    "qrcode": "^1.5.1",
+    "starknet": "3.18.2",
+    "svelte": "^3.50.1",
+    "svelte-check": "^2.7.1",
+    "svelte-preprocess": "^4.10.6",
+    "tslib": "^2.3.1",
+    "typescript": "^4.8.3",
+    "vite": "^3.0.4"
+  },
+  "type": "module"
 }
