+++ examples/get-starknet/package.json
@@ -10,7 +10,7 @@
   "devDependencies": {
     "@starknet/burner": "^0.3.11",
     "get-starknet": "^1.5.0",
-    "starknet": "^3.18.2",
+    "starknet": "3.18.2",
     "vite": "^3.0.9"
   }
 }
