+++ examples/starknet-react/package.json
@@ -34,6 +34,6 @@
     "@starknet-react/core": "^0.13.4",
     "@starknet/burner": "^0.3.12",
     "get-starknet": "^1.5.0",
-    "starknet": "^3.18.2"
+    "starknet": "3.18.2"
   }
 }
